warn only when the method is already defined
Including ActiveRecord::CommittedObserver was always displaying the
warning message because Module.included? is called after the modules
features have been appended to the class. This update checks for the
presence of after_commit before the features are appended to the
class--squelching the overzealous warnings. Defining after_commit after
including the module will continue to trigger the warning as expected.